:construction: refactor(ppddm-manager) for ARL mining processor.
